backend/requirements.txt
# API & Backend
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
sqlmodel>=0.0.16
pydantic>=2.0.0

# Core Scraping & AI
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx>=0.27.0
ollama>=0.4.0
google-generativeai>=0.8.0
openai>=1.50.0

# Document Processing
pdfplumber>=0.11.0
jinja2>=3.1.0
weasyprint>=62.0

# Data & Tracking
pandas>=2.2.0
openpyxl>=3.1.0

# Notifications
python-telegram-bot>=21.0

# CLI & UX
typer[all]>=0.12.0
rich>=13.7.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
python-docx>=1.1.0
